golang: Move from mod.rs -> parser.rs & add test cases
