+++ proxy-lib/src/http/firewall/rule/golang/mod.rs
@@ -207,79 +207,3 @@ impl RuleGolang {
         Ok(RequestAction::Allow(req))
     }
 }
-
-pub(super) struct GoPackage {
-    pub(super) fully_qualified_name: String,
-    pub(super) version: PragmaticSemver,
-}
-
-fn is_zip_download(path: &str) -> bool {
-    path.ends_with(".zip") && path.contains("/@v/")
-}
-
-/// Parses a Go module proxy zip URL path into a normalized module name and version.
-///
-/// Expected path shape: `/{module_path}/@v/{version}.zip`
-///
-/// Go's module proxy encodes uppercase letters as `!` + lowercase (e.g., `AikidoSec` →
-/// `!aikido!sec`, percent-encoded as `%21aikido%21sec`). We lowercase the whole path
-/// to normalise for malware-list lookup (which uses `LowerCaseEntryFormatter`).
-pub(super) fn parse_package_from_path(path: &str) -> Option<GoPackage> {
-    let path = path.trim_matches('/');
-
-    let (module_path_raw, rest) = path.split_once("/@v/")?;
-    let version_raw = rest.strip_suffix(".zip")?;
-
-    if module_path_raw.is_empty() || version_raw.is_empty() {
-        return None;
-    }
-
-    // Percent-decode the module path (handles %21 → !)
-    let module_path_decoded = percent_decode(module_path_raw);
-    // Lowercase for consistent malware-list lookup
-    let module_name = module_path_decoded.to_ascii_lowercase();
-
-    // Go versions always carry a 'v' prefix in the proxy URL; strip it for semver parsing
-    let version_str = version_raw.strip_prefix('v').unwrap_or(version_raw);
-    let version = PragmaticSemver::from_str(version_str)
-        .inspect_err(|err| {
-            tracing::debug!(
-                "failed to parse golang module ({module_name}) version (raw = {version_raw}): err = {err}"
-            );
-        })
-        .ok()?;
-
-    Some(GoPackage {
-        fully_qualified_name: module_name,
-        version,
-    })
-}
-
-/// Decodes percent-encoded characters in a URL path segment.
-/// Only handles the ASCII subset relevant to Go module paths (`%21` → `!`, etc.).
-fn percent_decode(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    let bytes = s.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] == b'%' && i + 2 < bytes.len() {
-            if let (Some(h), Some(l)) = (hex_val(bytes[i + 1]), hex_val(bytes[i + 2])) {
-                out.push((h << 4 | l) as char);
-                i += 3;
-                continue;
-            }
-        }
-        out.push(bytes[i] as char);
-        i += 1;
-    }
-    out
-}
-
-fn hex_val(b: u8) -> Option<u8> {
-    match b {
-        b'0'..=b'9' => Some(b - b'0'),
-        b'a'..=b'f' => Some(b - b'a' + 10),
-        b'A'..=b'F' => Some(b - b'A' + 10),
-        _ => None,
-    }
-}
